+++ main.tf
@@ -1,0 +1,10 @@
+terraform {
+  required_version = "~> 1.5.2"
+
+  required_providers {
+    aws = {
+      version = "~> 5.0.0"
+      source  = "hashicorp/aws"
+    }
+  }
+}