+++ variables.tf
@@ -11,7 +11,6 @@ variable "environment" {
 variable "aws_region" {
   type        = string
   description = "The AWS region that the app definition will read from, make sure this matches with the provider used for this module."
-  default     = "us-east-2"
 }
 
 variable "host_port" {
@@ -66,13 +65,48 @@ variable "ecs_svc_fargate_platform_version" {
 variable "ecs_svc_container_desired_count" {
   type        = number
   description = "Number of docker containers to run"
-  default     = 1
+  default     = 2
+}
+
+variable "ecs_svc_deployment_maximum_percent" {
+  type        = number
+  description = "Value representing the upper limit (as a percentage of the service's desiredCount) of the number of running tasks that can be running in a service during a deployment."
+  default     = 200
+}
+
+variable "ecs_svc_deployment_minimum_percent" {
+  type        = number
+  description = "Value representing the lower limit (as a percentage of the service's desiredCount) of the number of running tasks that must remain running and healthy in a service during a deployment."
+  default     = 100
+}
+
+variable "ecs_svc_force_new_deployment" {
+  type        = string
+  description = "Enable to force a new task deployment of the service. This can be used to update tasks to use a newer Docker image with same image/tag combination (e.g., myimage:latest), roll Fargate tasks onto a newer platform version, or immediately deploy ordered_placement_strategy and placement_constraints updates."
+  default     = false
+}
+
+variable "ecs_svc_capacity_provider_strategy" {
+  type = list(object({
+    capacity_provider = string
+    base              = number
+    weight            = number
+  }))
+  description = "Capacity provider strategies to use for the service. Can be one or more. These can be updated without destroying and recreating the service only if force_new_deployment = true and not changing from 0 capacity_provider_strategy blocks to greater than 0, or vice versa."
+  default     = []
+}
+
+
+variable "ecs_svc_enable_deployment_circuit_breaker" {
+  type        = bool
+  description = "Enable ECS deployment circuit breaker. This will enable ECS to roll back to the previous deployment if the new deployment fails."
+  default     = true
 }
 
 variable "ecs_svc_health_check_grace_period_seconds" {
   type        = number
   description = "Seconds to ignore failing load balancer health checks on newly instantiated tasks to prevent premature shutdown, up to 2147483647. Only valid for services configured to use load balancers."
-  default     = 30
+  default     = 5
 }
 
 variable "ecs_svc_enable_ssm" {
@@ -97,6 +131,18 @@ variable "ecs_task_network_mode" {
   default     = "awsvpc"
 }
 
+variable "ecs_svc_enable_deployment_event_alerts" {
+  type        = bool
+  description = "To enable or disable cloudwatch rule and sns topic creation for ECS deployment events."
+  default     = false
+}
+
+variable "ecs_svc_deployment_events" {
+  type        = set(string)
+  description = "List of ECS deployment events to send to the SNS topic."
+  default     = ["SERVICE_DEPLOYMENT_FAILED"]
+}
+
 variable "container_image_uri" {
   type        = string
   description = "Docker image to run in the ECS cluster"
@@ -117,6 +163,11 @@ variable "container_task_definition_protocol" {
   type        = string
   description = "The protocol that's used for the port mapping. Valid values are tcp and udp. The default is tcp."
   default     = "tcp"
+
+  validation {
+    condition     = can(regex("^(tcp|udp)$", var.container_task_definition_protocol))
+    error_message = "Must be either tcp or udp"
+  }
 }
 
 variable "container_environment_variables" {
@@ -131,7 +182,7 @@ variable "container_environment_variables" {
 variable "container_health_check_retries" {
   type        = number
   description = "The number of times to retry a failed health check before the container is considered unhealthy. You may specify between 1 and 10 retries."
-  default     = 3
+  default     = 2
 }
 
 variable "container_health_check_command" {
@@ -141,20 +192,20 @@ variable "container_health_check_command" {
 
 variable "container_health_check_timeout" {
   type        = number
-  description = "The period of time (in seconds) to wait for a health check to succeed before it's considered a failure. You may specify between 2 and 60 seconds. The default value is 5 seconds."
-  default     = 5
+  description = "The period of time (in seconds) to wait for a health check to succeed before it's considered a failure. You may specify between 2 and 60 seconds. The default value is 3 seconds."
+  default     = 3
 }
 
 variable "container_health_check_interval" {
   type        = number
-  description = "The period of time (in seconds) between each health check. You may specify between 5 and 300 seconds. The default value is 10 seconds."
-  default     = 10
+  description = "The period of time (in seconds) between each health check. You may specify between 5 and 300 seconds. The default value is 5 seconds."
+  default     = 5
 }
 
 variable "container_health_check_start_period" {
   type        = number
-  description = "The optional grace period to provide containers time to bootstrap in before failed health checks count towards the maximum number of retries. You can specify between 0 and 300 seconds. By default, startPeriod is disabled."
-  default     = 30
+  description = "The optional grace period to provide containers time to bootstrap in before failed health checks count towards the maximum number of retries. You can specify between 0 and 300 seconds."
+  default     = 10
 }
 
 ### IAM ###
@@ -187,6 +238,24 @@ variable "log_stream_prefix" {
   type        = string
   description = "Use the awslogs-stream-prefix option to associate a log stream with the specified prefix, the container name, and the ID of the Amazon ECS task that the container belongs to. "
   default     = "ecs"
+}
+
+### SNS ###
+variable "sns_topic_subscription_protocol" {
+  type        = string
+  description = "The protocol you want to use. Supported protocols include: [email, email-json, http, https, sqs, sms, lambda]"
+  default     = "https"
+
+  validation {
+    condition     = can(regex("^(email|email-json|http|https|sqs|sms|lambda)$", var.sns_topic_subscription_protocol))
+    error_message = "Must be either email, email-json, http, https, sqs, sms or lambda"
+  }
+}
+
+variable "sns_topic_subscription_endpoint" {
+  type        = string
+  description = "The endpoint that you want to receive notifications."
+  default     = ""
 }
 
 ### LOAD BALANCER ###
@@ -225,6 +294,12 @@ variable "alb_target_group_target_type" {
   default     = "ip"
 }
 
+variable "alb_target_group_deregistration_delay" {
+  type        = number
+  description = "Amount of time for Elastic Load Balancing to wait before changing the state of a deregistering target from draining to unused. The range is 0-3600 seconds. The default value is 300 seconds."
+  default     = 0
+}
+
 variable "alb_listener_action_type" {
   type        = string
   description = "Type of routing action. Valid values are [forward, redirect, fixed-response, authenticate-cognito and authenticate-oidc]"
@@ -244,14 +319,14 @@ variable "alb_listener_certificate_arn" {
 
 variable "alb_target_group_health_check_healthy_threshold" {
   type        = number
-  description = "Number of consecutive health check successes required before considering a target healthy. The range is 2-10. Defaults to 3."
-  default     = 3
+  description = "Number of consecutive health check successes required before considering a target healthy. The range is 2-10. Defaults to 2."
+  default     = 2
 }
 
 variable "alb_target_group_health_check_interval" {
   type        = number
-  description = "Approximate amount of time, in seconds, between health checks of an individual target. The range is 5-300. For lambda target groups, it needs to be greater than the timeout of the underlying lambda. Defaults to 30."
-  default     = 30
+  description = "Approximate amount of time, in seconds, between health checks of an individual target. The range is 5-300. For lambda target groups, it needs to be greater than the timeout of the underlying lambda. Defaults to 10."
+  default     = 10
 }
 
 variable "alb_target_group_health_check_matcher" {
@@ -263,13 +338,13 @@ variable "alb_target_group_health_check_matcher" {
 variable "alb_target_group_health_check_timeout" {
   type        = number
   description = "Amount of time, in seconds, during which no response from a target means a failed health check. The range is 2–120 seconds. "
-  default     = 15
+  default     = 5
 }
 
 variable "alb_target_group_health_check_unhealthy_threshold" {
   type        = number
-  description = "Number of consecutive health check failures required before considering a target unhealthy. The range is 2-10. Defaults to 3."
-  default     = 3
+  description = "Number of consecutive health check failures required before considering a target unhealthy. The range is 2-10. Defaults to 2."
+  default     = 2
 }
 
 variable "alb_stickiness_enabled" {